+++ src/modules/history.py
@@ -1,5 +1,6 @@
 import os
 import streamlit as st
+from streamlit_chat import message
 
 
 class ChatHistory:
@@ -10,14 +11,14 @@ class ChatHistory:
     def default_greeting(self):
         return "Hey ! 👋"
 
-    def default_prompt(self, thingy, topic):
-        return f'Hello! Ask me anything about the {thingy} "{topic}" 🤗'
+    def default_prompt(self, topic):
+        return f"Hello ! Ask me anything about {topic} 🤗"
 
     def initialize_user_history(self):
         st.session_state["user"] = [self.default_greeting()]
 
     def initialize_assistant_history(self, uploaded_file):
-        st.session_state["assistant"] = [self.default_prompt("document", uploaded_file.name)]
+        st.session_state["assistant"] = [self.default_prompt(uploaded_file.name)]
 
     def initialize(self, uploaded_file):
         if "assistant" not in st.session_state:
@@ -33,6 +34,18 @@ class ChatHistory:
 
     def append(self, mode, message):
         st.session_state[mode].append(message)
+
+    def generate_messages(self, container):
+        if st.session_state["assistant"]:
+            with container:
+                for i in range(len(st.session_state["assistant"])):
+                    message(
+                        st.session_state["user"][i],
+                        is_user=True,
+                        key=f"{i}_user",
+                        avatar_style="big-smile",
+                    )
+                    message(st.session_state["assistant"][i], key=str(i), avatar_style="thumbs")
 
     def load(self):
         if os.path.exists(self.history_file):
